+++ integrationtests/scenarios/i002520-interproject-references-constraint/before/A/ClassA.cs
@@ -1,0 +1,12 @@
+﻿using B;
+
+namespace A
+{
+	public class ClassA
+	{
+		public ClassA()
+		{
+			var classB = new ClassB();
+		}
+	}
+}