+++ integrationtests/scenarios/i002520-interproject-references-constraint/before/B/ClassB.cs
@@ -1,0 +1,6 @@
+﻿namespace B
+{
+	public class ClassB
+	{
+	}
+}