+++ src/Paket.Core/Installation/RestoreProcess.fs
@@ -193,17 +193,22 @@ let saveToFile newContent (targetFile:FileInfo) =
                     File.ReadAllText targetFile.FullName
                 else
                     ""
+            
+            let written =
+                if newContent <> oldContent then
+                    if verbose then
+                        tracefn " - %s created" targetFile.FullName
 
-            if newContent <> oldContent then
-                if verbose then
-                    tracefn " - %s created" targetFile.FullName
+                    if targetFile.Exists then
+                        File.SetAttributes(targetFile.FullName,IO.FileAttributes.Normal)
+                    File.WriteAllText(targetFile.FullName,newContent)
+                    true
+                else
+                    if verbose then
+                        tracefn " - %s already up-to-date" targetFile.FullName
+                    false
 
-                File.WriteAllText(targetFile.FullName,newContent)
-            else
-                if verbose then
-                    tracefn " - %s already up-to-date" targetFile.FullName
-
-            targetFile.FullName
+            written,targetFile.FullName
         with
         | exn when trials > 0 ->
             if verbose then
@@ -235,7 +240,7 @@ let extractRestoreTargets root =
     else
         let result = extractElement root "Paket.Restore.targets"
         copiedElements := true
-        result
+        result |> snd
 
 let CreateInstallModel(alternativeProjectRoot, root, groupName, sources, caches, force, package) =
     async {
@@ -317,13 +322,14 @@ let createPaketPropsFile (lockFile:LockFile) (cliTools:ResolvedPackage seq) (pac
                  yield! packageReferences
                  yield "    </ItemGroup>"])
             |> fun xs -> String.Join(Environment.NewLine,xs)
-                
+ 
+    // When updating the PaketPropsVersion be sure to update the Paket.Restore.targets which checks this value
     let content = 
         sprintf """<?xml version="1.0" encoding="utf-8" standalone="no"?>
 <Project ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
     <PropertyGroup>
         <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-        <PaketPropsVersion>5.174.2</PaketPropsVersion>
+        <PaketPropsVersion>5.185.3</PaketPropsVersion>
         <PaketPropsLoaded>true</PaketPropsLoaded>
     </PropertyGroup>
 %s
@@ -332,7 +338,7 @@ let createPaketPropsFile (lockFile:LockFile) (cliTools:ResolvedPackage seq) (pac
             cliParts
             packagesParts
 
-    saveToFile content fileInfo |> ignore
+    saveToFile content fileInfo
 
 let createPaketCLIToolsFile (cliTools:ResolvedPackage seq) (fileInfo:FileInfo) =
     if Seq.isEmpty cliTools then
@@ -471,7 +477,14 @@ let createProjectReferencesFiles (lockFile:LockFile) (projectFile:ProjectFile) (
     createPaketCLIToolsFile cliTools paketCLIToolsFileName
     
     let propsFile = ProjectFile.getPaketPropsFileInfo projectFileInfo
-    createPaketPropsFile lockFile cliTools packages propsFile
+    let written,_ = createPaketPropsFile lockFile cliTools packages propsFile
+    if written then
+        try
+            let fi = FileInfo(Path.Combine(projectFileInfo.Directory.FullName,"obj","project.assets.json"))
+            if fi.Exists then
+                fi.Delete()
+        with 
+        | _ -> ()
 
     // Write "cached" file, this way msbuild can check if the references file has changed.
     let paketCachedReferencesFileName = FileInfo(Path.Combine(projectFileInfo.Directory.FullName,"obj",projectFileInfo.Name + ".paket.references.cached"))
@@ -545,6 +558,20 @@ let private isRestoreUpDoDate (lockFileName:FileInfo) (lockFileContents:string) 
         let oldContents = File.ReadAllText(restoreCacheFile)
         oldContents = lockFileContents
     else false
+
+let internal WriteGitignore restoreCacheFile =
+    let folder = FileInfo(restoreCacheFile).Directory
+    let rec isGitManaged (folder:DirectoryInfo) =
+        if File.Exists(Path.Combine(folder.FullName, ".gitignore")) then true else
+        if isNull folder.Parent then false else
+        isGitManaged folder.Parent
+   
+    if isGitManaged folder.Parent then
+        let restoreCacheGitIgnoreFile = Path.Combine(folder.FullName, ".gitignore")
+        let contents = 
+            ".gitignore\npaket.restore.cached"
+            |> normalizeLineEndings
+        saveToFile contents (FileInfo restoreCacheGitIgnoreFile) |> ignore
 
 let IsRestoreUpToDate(lockFileName:FileInfo) =
     let newContents = File.ReadAllText(lockFileName.FullName)
@@ -688,5 +715,6 @@ let Restore(dependenciesFileName,projectFile,force,group,referencesFileNames,ign
                     CreateScriptsForGroups lockFile.Value groups
                     if isFullRestore then
                         let restoreCacheFile = Path.Combine(root, Constants.PaketRestoreHashFilePath)
-                        File.WriteAllText(restoreCacheFile, newContents))
+                        saveToFile newContents (FileInfo restoreCacheFile) |> ignore
+                        WriteGitignore restoreCacheFile)
             )